requirements.txt
pandas==1.5.2
xalpha==0.11.7
chinese-calendar==1.8.0
matplotlib==3.6.2
numpy==1.24.1
concurrent-log-handler==0.9.20
termcolor==2.1.1
selenium==4.7.2
retry==0.9.2
PyMySQL==1.0.2
SQLAlchemy==1.4.45
PyYAML==6.0
